move getmore connection test
